Move uiHelpers file
